Cosmetic changes for vulners plugin
